// generated/src/app/api-models/cap-report-setup.model.ts
/* Copyright (c) 2020 . All Rights Reserved. */

export class CapReportSetup {
  seqCprptId: number;
  jobId: string;
  requestUser: string;
  requestDate: Date;
  requestType: string;
  action: string;
  capReportTemplate: string;
  seqCcalcId: number;
  capEntityCode: string;
  seqCapPool1: number;
  seqCapPool2: number;
  template: string;
  comments: string;
  securityCode: string;
  insertDatetime: Date;
  insertUser: string;
  insertProcess: string;
  updateDatetime: Date;
  updateUser: string;
  updateProcess: string;
  inProcess: string;
}